Create per-course secrets for anonymizing students
Rather than use just the static `SECRET_KEY` as a hash key for
anonymizing student IDs, create at least one key per course.
This is easily extended to create more course keys depending on
the purpose: whether it's an ID needed for a third-party service
provider, like Qualtrics or SurveyMonkey, or an ID needed for
forum data (that, without anonymization, is easy to trace back to
a particular student and link up with the rest of the data).