WIP
moving things back because lib/ can't import from djangoapps/
committing here to save work.

so if continuing with this approach, either overload
anonymous_student_id as nate originally did (and hope for the best) or
implement a new filed on the ModuleSystem to carry around the
more_anonymous_student_id.

@dave has suggested that this is a perfect thing t odo with xblocks and
can be more cleanly implemented there. That's probably the next approach
to look at.